// src/types/Child.ts
export type Child = {
	childId: string;
	institutionId: string;
	groupId: string;
	createdTime: string;
	name: {
		fullName: string;
		firstName: string;
		middleName: string;
		lastName: string;
	};
	birthday: string | null;
	homeAddress: string | null;
	extraInfo: string;
	language: string;
	nationality: string;
	birthplace: string;
	gender: number;
	startDate: string;
	endDate: string | null;
	leavingReason: string | null;
	isTestChild: boolean;
	relations: any | null;
	image: {
		small: string;
		large: string;
		empty: boolean;
		colorCode: number;
	};
	isSleeping: boolean;
	naps: any[];
	hasVacation: boolean;
	isSick: boolean;
	isAbsent: boolean;
	leaves: any[];
	onBus: boolean;
	onTrip: boolean;
	statuses: any[];
	statusRegistrations: any[];
	checkins: Array<{
		childCheckinId: string;
		childId: string;
		institutionId: string;
		groupId: string;
		checkinTime: string;
		pickupTime: string | null;
		pickupRelationId: string;
		goHomeWithChildId: string;
		checkoutTime: string | null;
		checkinLoginId: string;
		checkoutLoginId: string;
		autoCheckedOut: boolean;
		deletedAt: string | null;
		hours: number;
		checkinStatements: any | null;
	}>;
	checkedIn: boolean;
	checkinTime: string | null;
	pickupTime: string | null;
	pickupRelationId: string | null;
	pickupName: string | null;
};